+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Refactor: Implement as startup script **needs Firebot v5.37.0 or newer**
 - …
 
 ## v0.1.3